+++ src/app.module.ts
@@ -18,7 +18,6 @@ import { PartnerModule } from './components/partner/partner.module';
 import { PartnershipModule } from './components/partnership/partnership.module';
 import { ProductModule } from './components/product/product.module';
 import { ProjectModule } from './components/project/project.module';
-import { RoleModule } from './components/role/role.module';
 import { ScriptureModule } from './components/scripture/scripture.module';
 import { SearchModule } from './components/search/search.module';
 import { SongModule } from './components/song/song.module';
@@ -56,7 +55,6 @@ import { CoreModule, LoggerModule } from './core';
     WorkflowModule,
     TimeZoneModule,
     PartnerModule,
-    RoleModule,
     FundingAccountModule,
   ],
   controllers: [],
