+++ src/components/authorization/authorization.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
-import { AuthorizationResolver } from './authorization.resolver';
-import { AuthorizationService } from './authorization.service';
+import { AuthorizationService } from '../authorization/authorization.service';
 
 @Module({
-  providers: [AuthorizationService, AuthorizationResolver],
+  providers: [AuthorizationService],
   exports: [AuthorizationService],
 })
 export class AuthorizationModule {}
