--- src/components/authorization/dto/add-property-to-security-group.dto.ts
@@ -1,17 +1,0 @@
-import { Field, InputType } from '@nestjs/graphql';
-import { IdField } from '../../../common';
-
-@InputType()
-export class AddPropertyToSecurityGroup {
-  @IdField()
-  readonly sgId: string;
-
-  @Field()
-  readonly property: string;
-}
-
-@InputType()
-export abstract class AddPropertyToSecurityGroupInput {
-  @Field()
-  readonly request: AddPropertyToSecurityGroup;
-}
