--- src/components/authorization/dto/create-security-group.dto.ts
@@ -1,22 +1,0 @@
-import { Field, InputType, ObjectType } from '@nestjs/graphql';
-import { IdField } from '../../../common';
-
-@InputType()
-export class CreateSecurityGroup {
-  @Field()
-  readonly name: string;
-}
-
-@InputType()
-export abstract class CreateSecurityGroupInput {
-  @Field()
-  readonly request: CreateSecurityGroup;
-}
-
-@ObjectType()
-export class CreateSecurityGroupOutput {
-  @Field()
-  success: boolean;
-  @IdField({ nullable: true })
-  id: string | null;
-}
