+++ src/components/authorization/dto/index.ts
