--- src/components/authorization/dto/security-group.dto.ts
@@ -1,11 +1,0 @@
-import { Field, ObjectType } from '@nestjs/graphql';
-import { IdField } from '../../../common';
-
-@ObjectType()
-export class SecurityGroup {
-  @IdField()
-  readonly id: string;
-
-  @Field()
-  readonly name: string;
-}
