+++ src/components/authorization/index.ts
@@ -1,1 +1,1 @@
-export * from './authorization.service';
+export * from './dto';
