+++ src/components/project/project-member/events/index.ts
@@ -1,0 +1,1 @@
+export * from './team-member-created.event';
