+++ src/components/project/project-member/events/team-member-created.event.ts
@@ -1,0 +1,4 @@
+// a team member was created and added to a project
+export class TeamMemberCreatedEvent {
+  constructor(readonly projectId: string, readonly userId: string) {}
+}
