+++ src/components/project/project-member/project-member.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
+import { AuthorizationModule } from '../../authorization/authorization.module';
 import { UserModule } from '../../user/user.module';
 import { ProjectMemberResolver } from './project-member.resolver';
 import { ProjectMemberService } from './project-member.service';
 
 @Module({
-  imports: [UserModule],
+  imports: [UserModule, AuthorizationModule],
   providers: [ProjectMemberResolver, ProjectMemberService],
   exports: [ProjectMemberService],
 })
