--- src/components/role/handlers/index.ts
@@ -1,1 +1,0 @@
-export * from './update-member-role.handler';
