--- src/components/role/index.ts
@@ -1,1 +1,0 @@
-export * from './dto';
