--- src/components/role/role.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import * as handlers from './handlers';
-
-@Module({
-  providers: [...Object.values(handlers)],
-})
-export class RoleModule {}
