+++ test/utility/index.ts
@@ -16,8 +16,8 @@ export * from './create-person';
 export * from './create-zone';
 export * from './create-workflow';
 export * from './create-state';
-export * from './create-security-group';
-export * from './create-permission';
+// export * from './create-security-group.ts';
+// export * from './create-permission';
 export * from './register';
 export * from './login';
 export * from './logout';
